fixes handling of union types
In the dataclasses conversion code type that is a member of a union
was not properly checked for if it was a member and so there would
always be an error.

For instance `FlyteFile.path` is `Union[str,Pathlike]` and so `str !=
Union[str,Pathlike]`.

This patch adds support for checking that a type is part of a union
and a satisfactory type.

Signed-off-by: Samuel Lotz <sam@examol.io>